refactor!: Consistent app template/Xcode project name (#1485)
* refactor!(project): Use consistent project name

* chore(deps): Remove unneeded xmlescape dependency

* fix(build): Consistent handling of archivePath flags

* feat!: Pull in CordovaLib with SwiftPM

* fix(orientation): Set orientation values in Xcodeproj

Otherwise when Xcode consolidates the plist file it will probably
overwrite the ones in the plist with the ones defined in the Xcodeproj
file (which is where it wants to put things by default nowadays).

* fix(pbxproj): Fix compat with some plugins

Some plugins look up the root group of the Xcode project based on the
name "CustomTemplate", which isn't really an ideal thing to do but I'm
unsure if there are better options and this name doesn't actually get
displayed anywhere so we might as well keep the label for compatibility
reasons.

* fix(project): Validate that App.xcodeproj exists

This is a guard against running a newer version of Cordova iOS tools
against an older project with a different name (which will fail due to
now-hardcoded assumptions about the project name).

* fix(run): Launch app based on the product name

+++ tests/spec/createAndBuild.spec.js
@@ -36,10 +36,9 @@ const makeTempDir = () => path.join(
  * @param {String} projectName
  */
 function verifyProjectFiles (tmpDir, projectName) {
-    expect(fs.existsSync(path.join(tmpDir, projectName))).toBe(true);
-    expect(fs.existsSync(path.join(tmpDir, `${projectName}.xcodeproj`))).toBe(true);
-    expect(fs.existsSync(path.join(tmpDir, `${projectName}.xcworkspace`))).toBe(true);
-    expect(fs.existsSync(path.join(tmpDir, 'CordovaLib'))).toBe(true);
+    expect(fs.existsSync(path.join(tmpDir, 'App'))).toBe(true);
+    expect(fs.existsSync(path.join(tmpDir, 'App.xcodeproj'))).toBe(true);
+    expect(fs.existsSync(path.join(tmpDir, 'App.xcworkspace'))).toBe(true);
 }
 
 /**
@@ -50,11 +49,15 @@ function verifyProjectFiles (tmpDir, projectName) {
  * @param {String} expectedBundleIdentifier
  */
 function verifyProjectBundleIdentifier (tmpDir, projectName, expectedBundleIdentifier) {
-    const pbxproj = path.join(tmpDir, `${projectName}.xcodeproj`, 'project.pbxproj');
+    const pbxproj = path.join(tmpDir, 'App.xcodeproj', 'project.pbxproj');
     const xcodeproj = xcode.project(pbxproj);
     xcodeproj.parseSync();
-    const actualBundleIdentifier = xcodeproj.getBuildProperty('PRODUCT_BUNDLE_IDENTIFIER');
+
+    const actualBundleIdentifier = xcodeproj.getBuildProperty('PRODUCT_BUNDLE_IDENTIFIER', undefined, 'App');
     expect(actualBundleIdentifier).toBe(`"${expectedBundleIdentifier}"`);
+
+    const actualBundleName = xcodeproj.getBuildProperty('PRODUCT_NAME', undefined, 'App');
+    expect(actualBundleName).toBe(`"${projectName}"`);
 }
 
 /**
@@ -74,7 +77,7 @@ function verifyBuild (tmpDir) {
 
     const Api = require(path.join(tmpDir, 'cordova', 'Api.js'));
 
-    return expectAsync(new Api('ios', tmpDir).build({ emulator: true }))
+    return expectAsync(new Api('ios', tmpDir).build({ emulator: true, buildFlag: ['-quiet'] }))
         .toBeResolved();
 }
 
@@ -93,26 +96,28 @@ async function verifyCreateAndBuild (tmpDir, packageName, projectName) {
         .then(() => verifyBuild(tmpDir));
 }
 
-describe('create', () => {
-    let tmpDir;
+if (process.platform === 'darwin') {
+    describe('create', () => {
+        let tmpDir;
 
-    beforeEach(function () {
-        tmpDir = makeTempDir();
+        beforeEach(function () {
+            tmpDir = makeTempDir();
+        });
+
+        afterEach(() => {
+            fs.rmSync(tmpDir, { recursive: true, force: true });
+        });
+
+        it('Test#001 : create project with ascii name, no spaces', () => {
+            const packageName = 'com.test.app1';
+            const projectName = 'testcreate';
+            return verifyCreateAndBuild(tmpDir, packageName, projectName);
+        }, 10 * 60 * 1000); // first build takes longer (probably cold caches)
+
+        it('Test#002 : create project with complicated name', () => {
+            const packageName = 'com.test.app2';
+            const projectName = '応応応応 hello & إثرا 用用用用';
+            return verifyCreateAndBuild(tmpDir, packageName, projectName);
+        }, 5 * 60 * 1000);
     });
-
-    afterEach(() => {
-        fs.rmSync(tmpDir, { recursive: true, force: true });
-    });
-
-    it('Test#001 : create project with ascii name, no spaces', () => {
-        const packageName = 'com.test.app1';
-        const projectName = 'testcreate';
-        return verifyCreateAndBuild(tmpDir, packageName, projectName);
-    }, 10 * 60 * 1000); // first build takes longer (probably cold caches)
-
-    it('Test#002 : create project with complicated name', () => {
-        const packageName = 'com.test.app2';
-        const projectName = '応応応応 hello & إثرا 用用用用';
-        return verifyCreateAndBuild(tmpDir, packageName, projectName);
-    }, 5 * 60 * 1000);
-});
+}
